modulegraph: adjust _scan_bytecode for new LOAD_ opcodes
Adjust the asserts in modulegraph's `_scan_bytecode` function
to account for new `LOAD_CONST_IMMORTAL` and `LOAD_SMALL_INT`
opcodes being used in addition or instead of `LOAD_CONST` to set
up arguments for the `IMPORT_NAME` call. The new opcodes were
introduced in python 3.14.0a2.

At least in theory, any of those three can be used to load the
`flags` value; and `LOAD_CONST` or `LOAD_CONST_IMMORTAL` can be used
to load the `names`.