tests: reproducibility: add a throw-away "calibration" build step
Add a throw-way "calibration" build step to both tests, in an attempt
to mitigate issues with non-reproducibility of byte-code compilation
result when repeatedly compiling the same module within the same
python process (e.g., within the context of the `pytest` runner).
For some modules, the very first byte-compilation result might differ
from all subsequent ones, so by putting the potential first compilation
into a throw-away step, we ensure that the tests work as expected
regardless of the order in which they are executed (or when executed
on their own).