hookutils: get_installer: treat the passed name as dist name first
Modify the `get_installer` hook utility function to treat the passed
name as dist name first and perform direct metadata lookup.

Only if that fails (matching dist is not found), treat the passed
name as importable package/module name, and try to look-up matching
dist via `importlib_metadata.packages_distributions()` mapping.
In contrast to previous implementation that considered only the
first matching dist (regardless of whether it had an `INSTALLER`
file or not), we now take the contents of first dist that has
`INSTALLER` file available, if any.

Finally, remove the macOS-specific fallback that tries to use
macports' `port` command to determine whether importable file was
installed via macports, and checks for "Cellar" in its path to
determine whether file was installed via homebrew. We do not require
this functionality in any of the few hooks that use this helper;
plus, homebrew-installed python packages should have `INSTALLER`
file that contains `brew` as installer identifier.