hookutils: setuptools: fix detection of vendored modules
Fix detection of `setuptools`-vendored modules (i.e., not packages)
in the `SetuptoolsInfo` hook utility class; for example, the
`setuptools/_vendor/typing_extensions.py` module.

When determining whether a package/module is vendored by `setuptools`,
we need to check if `setuptools` vendor directory path is in parents
of the candidate path, and also if the candidate path itself is
the same as the vendor directory path. The first check covers vendored
packages, but the second check is required to correctly handle
vendored modules.