tests: test_global_functions: skip part of test with 300-argument function
Starting with python >= 3.14.0a5, function calls with more than 30
arguments end up constructing a list with a sequence of BUILD_LIST,
LOAD*, and LIST_APPEND opcodes; it then converts the list using
CALL_INTRINSIC_1 opcode with INTRINSIC_LIST_TO_TUPLE, and calls
the function with CALL_FUNCTION_EX opcode.

We currently have no use case for such lists, so instead of trying to
accommodate this in our bytecode scanner, skip the corresponding part
of the test when running under python >= 3.14.