build: fix handling of directories captured by glob in --add-data
Fix handling of directories captured by glob in the
`format_binaries_and_datas` helpers; in contrast to the case when
the given source path is a directory and we collect the contents
of that directory into the specified target directory, when the
given source path contains a glob that matches directories (in
addition or instead of files), those directories themselves should
be collected into the specified target directory (as opposed to
their content being collected into the specified target directory).

This aims to ensure that `--add-data data_dir:data_dir` and
`--add-data data_dir/*:data_dir` end up behaving in the same way
when `data_dir` contains sub-directories.