bootloader: perform AIX-specific adjustment of python lib name at build time
On AIX, the python shared library might in fact be an ar archive with
 shared object inside it, and needs to be `dlopen`'ed with full name
(for example, `libpython3.9.a(libpython3.9.so)` and with `RTLD_MEMBER`
flag set.

Up until now, we had bootloader check if the python shared library's
name is .a, and change the name into `libpython3.X.a(libpython3.X.so)`
if necessary. However, it makes more sense to perform this check during
the build, and store the final (adjusted as necessary) python library
name for bootloader to use.

This eliminates the corresponding AIX-specific codepath in bootloader,
where the only remaining requirement is that we pass `RTLD_MEMBER` flag
to `dlopen`; and as it seems that we can use this flag even with
non-archived shared libraries (i.e., "regular" .so files), we can
always have it set.

Having the name adjustment implemented in the building code also
ensures that if ever needed, more complex form of library name
resolution can be implemented; for example, parsing the archive's
header to dynamically infer the name of the shared object, instead
of assuming fixed name format.