splash: fix names of Tcl and Tk shared libraries under AIX
Similar to the Python shared library, the Tcl and Tk shared library on
AIX might be shared objects stored in an .a archive, and need to be
`dlopen`'ed using their full name; for example, `libtcl.a(libtcl.so.8.6)`
and `libtk.a(libtk.so.8.6)`.

So have the splash build target stage fully resolve the names of
shared libraries and store the resolved names for bootloader to use.