bootloader: waf: update the strip feature/tool
Update the strip feature/tool based on the latest version of the
example from waf's playground:
https://gitlab.com/ita1024/waf/-/blob/master/playground/strip/strip_hack.py

In particular, the updated version checks whether the preceding
link task was ran or not, and based on its status decides whether
to run strip task or not. This prevents the strip task from being
ran twice per binary (once as part of `build_*` command and then
again as part of `install_*` command), which not only looks odd
in the build log, bit might also result in error with `strip`
implementations that do not allow being ran against already-stripped
binary (for example, the `strip` utility on AIX).

Change the displayed keyword to "Stripping binary", and change the
binary color to YELLOW, to match the color used in the preceding
link command.