bootloader: waf: use -X32_64 flag with strip on AIX
On AIX, the `strip` utility needs to be explicitly told whether to
process 32-bit object files (`-X32`), 64-bit object files (`-X64`) or
both (`-X32_64`).

The default is to process 32-bit object files (ignore 64-bit objects).
The mode can be controlled via `OBJECT_MODE` environment variable, and
explicitly overridden via `-X` option. Therefore, use `-X32_64` to
accommodate both 32-bit and 64-bit case.

Fixes the following error when building bootloader on AIX:

```
strip: run_d -- 0654-423 run_d is not valid in the current object file mode.
       Use the -X option to specify the desired object mode.
```

when building with `/opt/freeware/bin/gcc` and `--target-arch=64bit`
passed to `waf`, and without `OBJECT_MODE` environment variable set
(which is not honored by said compiler, so `--target-arch` is needed).

A long, long time ago (PyInstaller < 4.0), we used to set `-X32` and
`-X64` flag depending on the target architecture, but this was then
removed as part of #4731 in an attempt to mitigate #4211, and with
argument  that the flag is redundant anyway, since all AIX utilities
honor `OBJECT_MODE`.

However, this is/was the case with IBM's `xlc` compiler, but not
with `/opt/freeware/gcc` that is nowadays available as part of AIX
open-source toolbox. In addition, the true cause of errors in #4211
was that `xlc_r` executable was being used in lieu of `strip`
executable, and was therefore raising error about unknown arguments.
This, in turn, was caused by bug in the code that was at the time
used to determine the name  of the `strip` utility based on the name
of compiler executable.