bootloader: waf: on AIX, use static build of bundled zlib copy
On AIX, force `static_zlib=True` so that bundled copy of zlib is
statically built and used, without having to explicitly specify
the `--static-zlib` option.

Remove the AIX-specific block that removed 'Z' from `libs` and added
'z' to `staticlibs`; this seems to be a remnant of older times where
we built against system-provided static zlib library. But there seems
to be little advantage in trying to use that over the bundled zlib
sources; plus, it requires the build system to provide development
headers and static copy of the library.