building: on AIX, suppress warnings about strip failing on already-stripped binaries
On AIX, the `strip` utility returns 255 and displays the following error
message when used with already-stripped binary:

```
strip: <name> -- 0654-419 The specified archive file was already stripped.
```

Catch these errors based on what appears to be the error code (0654-419;
the rest of the message can presumably be localized), and suppress
the corresponding warning messages to avoid polluting build log.