bootloader: set _PYI_APPLICATION_HOME_DIR immediately after creating temporary directory
Have the bootloader in onefile parent codepath set the
`_PYI_APPLICATION_HOME_DIR` environment variable immediately after
resolving and creating temporary directory.

This improves localization of corresponding `setenv` and `getenv`
calls (i.e., we set the environment variable in one branch, and
retrieve it in the other branch of the same if-block).