ci: enable basic CI with free-threading python 3.13
We cannot run the full CI because not all packages in
`requirements-libraries.txt` provide free-threaded builds of their
wheels yet, and there is no way to filter them out in requirements
file.

So we keep treating 3.13t same as "-dev" version (which is better
than nothing), and with 3.14t, we should be able to use the
`sys_abi_features` environment marker from PEP 780 to skip just
the problematic packages.