tests: remove individual pytest.mark.timeout() marks
Remove individual `pytest.mark.timeouts()` marks from the few tests
that have them (`multiprocessing` and `Django` tests).

These individual marks are likely remnants of times when we had no
global timeout setting, or when the global setting was more restrictive
than individual ones; however, the global setting has been 10 minutes
for a while now, and these individual settings, which are more
restrictive, are unnecessary.

And in the case of multiprocessing tests, the 1-minute timeouts may
be too restrictive when running tests in parallel and when using the
free-threaded python, as CPU congestion might lead to tests being
slowed down.