bootloader: disable Windows error reporting dialog only in error codepath
Move the `SetErrorMode()` call that disables error reporting dialog
on Windows to the error-handling codepath that follows the call to
`Py_InitializeFromConfig()`.

When we were still using `Py_Initialize()`, we had to disable
error reporting before calling that function; because, as noted
in the original comment, it might have ended up calling `abort()`
on error. However, this is not the case with `Py_InitializeFromConfig()`,
which signals errors through the returned status structure. However,
the subsequent `PyStatus_Exception(status)` call in the error-handling
codepath might call `abort()`. So we need to disable error reporting
before calling that function; and because `PyStatus_Exception`
calls either `exit()` or `abort()`, we do not need to worry about
re-enabling it.

Formerly, we disabled error reporting only in debug builds; this is
somewhat counter-intuitive, but was done because debug bootloader
is used on CI and we did not want the error reporting dialog to
block the CI pipeline. However, it makes more sense to disable this
for both debug and release bootloader builds.