rthooks: pkg_resources: suppress deprecation warning during import
Have the run-time hook for `pkg_resources` explicitly suppress the
deprecation warning that is emitted with `setuptools` >= 80.9.0.

This silences the said warning in cases when `pkg_resources` ends
up collected into frozen application due to reference in some 3rd
party, where it may be referred in a codepath that is normally
not used. Instead of having to discover all such packages and
explicitly exclude `pkg_resources` via hooks, just silence the
warning - as far as the run-time hook is concerned, the issue with
`pkg_resources` being removed in the near future will solve itself
once the removal really happens.