ci: fix the Termux CI flow
Something changed in the latest version of Termux (or its packages),
and now installation of `psutil` fails due to `getifaddrs()` and
`freeifaddrs()` functions not being exported from the system
`ifaddrs.h` header.

This, in turn, seems to be caused by `pip` trying to compile the
`psutil` extensions with `x86_64-linux-android-clang`, which does
not define `__ANDROID_API__` and `__ANDROID_MIN_SDK_VERSION__`.
On the other hand, "regular" `clang` and `gcc` (which are symlinked
to the same underlying executable) define those, and can compile
`psutil`.

Therefore, explicitly set `CC=clang` at the start of the workflow.