bindepend: rework python shared library search
Rework the python shared library search in order to minimize the
amount of guesswork.

On Windows, use `sys.dllhandle` to obtain handle to the loaded
python DLL, and obtain its full path via `GetModuleFileName`
win32 API. This covers python.org Windows builds, Anaconda on
Windows, and MSYS2 python.

On other platforms, first explicitly check that python was built
with shared library support, by looking at `Py_ENABLE_SHARED`
variable in `sysconfig` module. On macOS, this is set to 0 for
.framework bundle builds, so also check the `PYTHONFRAMEWORK`
variable, which should be non-empty in this case. If shared library
(of .framework bundle) was not enabled, raise an error.

Obtain the name of python shared library from `INSTSONAME` variable
in `sysconfig` module. This should handle all variability across
different platforms, and also different build options (i.e., the
'd' suffix for debug builds, 't' for free-threading in python >= 3.13,
or both).

Remove all the related variables from the `PyInstaller.compat` module,
as they are not needed anymore.