bindepend: fix new shared library search for POSIX Anaconda builds
Turns out that while Anaconda Python for Linux and macOS does
provide Python shared library, it is built separately from the
interpreter executable. Therefore, the `sysconfig` module imported
from the interpreter executable reports `Py_ENABLE_SHARED` to be
0, and `INSTSONAME` contains the static library name.

This is in contrast to Debian-packaged Python and
`astral-sh/python-build-standalone` Linux/macOS builds, where the
interpreter executable is also statically linked, but `Py_ENABLE_SHARED`
is reported to be 1 and `INSTSONAME` contains the shared library name.

Therefore, we need to re-introduce some of the old guess-work code
to accommodate Anaconda Python on Linux and macOS, but limit it
only to this case.