building: revise the co_filename anonymization logic
Revise the logic for anonymization of `co_filename` in collected
code objects. Instead of trying to remove known path prefixes
from the current `co_filename`, we now construct the target
`co_filename` based on the code object's destination container
and destination name within it.

This makes the anonymization compliant with the requirements of
the test added in the preceding commit, with the following main
benefits:

- anonymization should now work for PyInstaller modules and
  scripts (bootstrap modules, run-time hooks, `pyi_splash`
  module, and `_pyi_rth_utils` supporting package) even when
  PyInstaller is installed in editable mode

- on case-insensitive operating systems (Windows), we side-step the
  issue of case mismatches

- we side-step the issues with vendored packages whose locations
  are dynamically added to `sys.path`; for example, the `co_filename`
  of `setuptools._vendor.jaraco.text` is now replaced with
  `setuptools/_vendor/jaraco/text/__init__.py` instead of
  `jaraco/text/__init__.py`.

The last point is important if we ever want to have our `PyiFrozenLoader`
dynamically preprend `sys._MEIPASS` to `co_filename` on code objects
it loads from the PYZ archive (for example, to accommodate code that
tries to infer the location of code object's module from its `co_filename`).