building: avoid creating dummy code objects for PEP-420 namespace packages
Avoid creating dummy ("empty") code objects for PEP-420 namespace
packages, as these are not used anywhere. We do need a TOC entry
in PYZ for our run-time module finder/loader to be aware of the
namespace package, but we can just as well create an entry with
data size set to 0.