setup: Drop setuptools, move to hatchling
Setuptools is forcing the migration to the new PEP 639 license
expressions by deprecating the old license field. However, the oldest
setuptools version that supports the new field is 77.0 and Python 3.8
support was dropped in 76.1.0. setuptools refuses to build if it sees
any configuration that it's too old to know about so we’re actually
incapable of using the new field without breaking Python 3.8, yet we’re
also being forced to.

I've been eyeing up switching to hatchling (just the unobtrusive build
backend, not Hatch the full-blown workflow manager) for a while and the
above finally made me try it. Not only does hatchling preserve backwards
compatibility for the old license field but it's also forgiving of
fields that it's too old to recognise so, if we did switch to the new
form of license expressions, the only damage would be that people
installing from source/sdist on Python 3.8 would be missing a
License-Expression field in pyinstaller-x.y.z.dist-info/METADATA.

Unfortunately though, in a misguided attempt to spur on the transition,
PEP 639 made license classifiers mutually exclusive to license
expressions. I don't want to have to break
pyinstaller/pyinstaller-hooks-contrib#637 so ironically, this means that
I'd prefer to still boycott the new license expressions.

I've tried to match the old setup and logic to select bootloaders/wheel
targets 1:1. The only change that should be visible from the outside is
that I added a few extra project URLs. I think that hatchling will be
much less maintenance than setuptools in the long run.