Remove empty pytest.ini allegedly needed by test entrypoints
This file was unintentionally being left behind by setuptools wheel
packaging and is now effectively being reintroduced by the switch to
hatchling. It looks like it was only used in a misguided way to prevent
the hook-sample test not use the root pytest.ini.

It's unclear what re-adding this file could do so I reckon
it's safer to preserve the old unintended behaviour by removing it.