hooks: add pre-safe-import-module hooks for `backports` and `jaraco`
Add pre-safe-import-module hooks for `backports` and `jaraco`, in
order to properly handle `setuptools`-vendored copies of these
top-level namespace (or namespace-like) packages in situations
when the `setuptools` vendored-package directory is not in the
python's search path at the start of analysis process.

Both packages have namespace-package semantics, and can therefore
be split between multiple locations (i.e., mixture of components
in regular site-packages and vendored components in `setuptools`
vendor directory); this requires additional considerations compared
to other vendored packages and modules (for which we have already
set up hooks). If such a top-level namespace package is *fully
vendored* (i.e., is provided solely by `setuptools`), we need to
set up an alias (which prevents creation of unnecessary top-level
package in PYZ archive). However, if it is *partially vendored*,
we need to extend the search paths for the package, to accommodate
corner case where `setuptools` vendored-package directory is not
in the search path at the start of analysis process.